add fix for imports embedded in scaladoc
Fixed issue where if some scaladoc was found before an entity such as
a class, trait or object, then the import declaration was inserted
into the second line of the scaladoc.

The issue occurred because we didn't take multiline scaladoc into
consideration. Once the first line of the scaladoc declaration
was processed, Scuggest incorrectly assumed that the next line would
be non-scaladoc.

The processing order has now changed to only skip over package and
package object declarations and empty lines. After that, the import
is inserted.

Also added some tests around this functionality and extracted the
source into inserts.py